+++ src/de.sormuras.bach/main/java/de/sormuras/bach/tool/Javadoc.java
@@ -18,13 +18,9 @@
 package de.sormuras.bach.tool;
 
 import java.util.List;
-import java.util.function.UnaryOperator;
 
 /** A {@code javadoc} call configuration. */
 public final class Javadoc implements WithModuleSourceOptionsCall<Javadoc> {
-
-  @FunctionalInterface
-  public interface Tweak extends UnaryOperator<Javadoc> {}
 
   private final List<Argument> arguments;